tools/submodules/cookeys/requirements.txt
colorama
pypiwin32
pycryptodome